refactor: Controller 인터페이스 패키지 위치 이동
